WIP for metadata section block
