+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { EditMetadataBlock } from './manage/Blocks/Metadata/Edit';
 export { ViewMetadataBlock } from './manage/Blocks/Metadata/View';
 
-export EditMetadataSectionBlock from './manage/Blocks/MetadataSection/Edit';
-export ViewMetadataSectionBlock from './manage/Blocks/MetadataSection/View';
+export EditMetadataSectionBlock from './manage/Blocks/MetadataSection/EditMetadataSection';
+export ViewMetadataSectionBlock from './manage/Blocks/MetadataSection/ViewMetadataSection';
